requirements.txt
streamlit==1.30.0
openai==1.12.0
google-generativeai==0.3.2
python-dotenv==1.0.0
psycopg2-binary==2.9.9
elevenlabs==0.2.28
pandas==2.1.4
plotly==5.18.0
numpy==1.26.3
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pydantic==2.5.3
pillow==10.2.0
pygame==2.5.2
PyYAML>=6.0
requests>=2.28.0
